--- src/coreMethods/cardToken/type.ts
@@ -1,6 +1,0 @@
-export type TFieldsCardTokenParams = {
-  cardId?: string;
-  cardholderName?: string;
-  identificationType?: string;
-  identificationNumber?: string;
-};
