+++ src/coreMethods/cardToken/types.ts
@@ -1,0 +1,10 @@
+export type TCardTokenParams = {
+  cardNumber?: string;
+  cardholderName?: string
+  identificationType?: string
+  identificationNumber?: string
+  securityCode?: string
+  cardExpirationMonth?: string
+  cardExpirationYear?: string
+  cardId?: string
+}
