+++ src/coreMethods/createCardToken/index.ts
@@ -1,14 +1,14 @@
 import { MercadoPagoInstance } from "../../mercadoPago/initMercadoPago";
-import type { TCardTokenParams } from "./types";
+import { TFieldsCardTokenParams } from "./types";
 
 /**
  * Token creation method.
  * 
- * @see {@link https://github.com/mercadopago/sdk-js/blob/main/API/core-methods.md#mp-instancecreatecardtokencardtokenparams method documentation}.
+ * @see {@link https://github.com/mercadopago/sdk-js/blob/main/API/fields.md#mp-instancefieldscreatecardtokennonpcidata method documentation}.
  */
-const createCardToken = async (cardTokenParams: TCardTokenParams) => {
+const createCardToken = async (cardTokenParams: TFieldsCardTokenParams) => {
   const instanceMercadoPago = await MercadoPagoInstance.init();
-  return instanceMercadoPago?.createCardToken(cardTokenParams);
+  return instanceMercadoPago?.fields.createCardToken(cardTokenParams);
 };
 
 export default createCardToken;
